requirements.txt
# Core dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0

# Video processing
ffmpeg-python>=0.2.0
opencv-python>=4.8.0
pillow>=10.0.0

# AI models
openai-whisper>=20231117
openai>=1.0.0
torch>=2.0.0
torchaudio>=2.0.0

# Database
sqlite3  # Built-in with Python
supabase>=2.0.0  # Optional for cloud sync

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0

# Utilities
pandas>=2.0.0
numpy>=1.24.0
python-magic>=0.4.27
tqdm>=4.66.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
pre-commit>=3.5.0 